[#96] refactor: AddLinkView 구조 분리
